pyproject.toml
[project]
name = "quantum-examples"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Raul Bezares", email = "rbezares5@gmail.com" }
]
dependencies = [
    "numpy>=1.24.4",
    "dwave-neal>=0.6.0",
    "pyqubo>=1.3.0",
    "dwave-system>=1.23.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "quantum_examples:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.2.2",
    "sphinx>=7.1.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/quantum_examples"]

[tool.ruff]
line-length = 120
select = ["ALL"]

[tool.ruff.pydocstyle]
convention = "google"
